core-service/pyproject.toml
[project]
name = "core-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiobotocore>=3.0.0",
    "confluent-kafka>=2.12.2",
    "fastapi[standard]>=0.127.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg[binary,pool]>=3.3.2",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy[asyncio]>=2.0.45",
]
